create a folder to test routes
